% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mbbs_data.R
\name{add_route_info}
\alias{add_route_info}
\title{Add route info to data scraped from old MBBS site}
\usage{
add_route_info(dt, county)
}
\arguments{
\item{dt}{a data.frame}

\item{county}{the name of the county to merge}
}
\description{
Add route info to data scraped from old MBBS site
}
