% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{add_sci_name}
\alias{add_sci_name}
\title{Add scientific name}
\usage{
add_sci_name(df, taxonomy)
}
\arguments{
\item{df}{a dataset with a `common_name` field}

\item{taxonomy}{data.frame from `get_ebird_taxonomy`}
}
\description{
Add scientific name
}
