% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{check_duplicate_submissions}
\alias{check_duplicate_submissions}
\title{Check that the submissions have the correct number of stops.}
\usage{
check_duplicate_submissions(dt, year = NULL)
}
\arguments{
\item{dt}{a `data.frame` imported using `import_ebird_data`}

\item{year}{a optional year to filter to}
}
\value{
a `data.frame` with observations that have > 1 submission for a route
  in  a year
}
\description{
Check that the submissions have the correct number of stops.
}
