% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{check_number_of_stops}
\alias{check_number_of_stops}
\title{Check that the submissions have the correct number of stops.}
\usage{
check_number_of_stops(dt)
}
\arguments{
\item{dt}{a `data.frame` imported using `import_ebird_data`}
}
\value{
a `data.frame` with observations that have a number of stops *other*
  than 1 or 20
}
\description{
Check that the submissions have the correct number of stops.
}
