% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_comments.R
\name{clean_comments}
\alias{clean_comments}
\title{Clean up the eBird comments}
\usage{
clean_comments(comments)
}
\arguments{
\item{comments}{a `character` vector of eBird comments}
}
\description{
Clean up the eBird comments
}
