% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mbbs_data.R
\name{clean_common_names}
\alias{clean_common_names}
\title{Clean the species names of the data scraped from the old MBBS website}
\usage{
clean_common_names(dt)
}
\arguments{
\item{dt}{a data.frame}
}
\description{
Clean the species names of the data scraped from the old MBBS website
}
