% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mbbs_data.R
\name{combine_site_ebird}
\alias{combine_site_ebird}
\title{Combine MBBS site data with eBird data}
\usage{
combine_site_ebird(x, at_year = 2010)
}
\arguments{
\item{x}{the output of `prepare_mbbs_data`}

\item{at_year}{the year which to begin using ebird data}
}
\description{
Combine MBBS site data with eBird data
}
