% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_comments.R
\docType{data}
\name{comment_spec}
\alias{comment_spec}
\title{Parser specification for ebird checklist comments}
\format{
An object of class \code{list} of length 5.
}
\usage{
comment_spec
}
\description{
Parser specification for ebird checklist comments
}
\keyword{datasets}
