% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_comments.R
\name{comment_workflow}
\alias{comment_workflow}
\title{Full workflow for processing comments}
\usage{
comment_workflow(ebird)
}
\arguments{
\item{ebird}{a `data.frame` of imported eBird counts}
}
\value{
a `data.frame` with one row per submission ID in `ebird`
}
\description{
Full workflow for processing comments
}
