% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{compute_ebird_counts}
\alias{compute_ebird_counts}
\title{Transformation of the ebird data to count data}
\usage{
compute_ebird_counts(ebird)
}
\description{
Transformation of the ebird data to count data
}
