% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{config}
\alias{config}
\title{Defines various configuration settings used in this project}
\format{
An object of class \code{list} of length 17.
}
\usage{
config
}
\description{
Including:
* filepaths/names
* regex patterns
* directories
}
\keyword{datasets}
