% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{conform_taxonomy}
\alias{conform_taxonomy}
\title{Conform taxonomy of MBBS data from different sources}
\usage{
conform_taxonomy(df, taxonomy)
}
\arguments{
\item{df}{a dataset with a `common_name` field}

\item{taxonomy}{data.frame from `get_ebird_taxonomy`}
}
\description{
Conform taxonomy of MBBS data from different sources
}
