% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-list.R
\name{convert_based_on_observer_table}
\alias{convert_based_on_observer_table}
\title{Take df, convert columns obs1/obs2/obs3 based on the
observer_table and then return the df with corrected columns}
\usage{
convert_based_on_observer_table(ebird_surveys, config)
}
\description{
Take df, convert columns obs1/obs2/obs3 based on the
observer_table and then return the df with corrected columns
}
