% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbbs.R
\name{create_route_level_0}
\alias{create_route_level_0}
\title{Create the route level dataset}
\usage{
create_route_level_0(ebird, stop_level_data, taxonomy, config = config)
}
\description{
Create the route level dataset
}
