% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbbs.R
\name{create_route_level_counts}
\alias{create_route_level_counts}
\title{Main function to create route level count dataset}
\usage{
create_route_level_counts(ebird, stop_level_data, taxonomy, config = config)
}
\description{
Main function to create route level count dataset
}
