% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbbs.R
\name{create_route_level_counts_0}
\alias{create_route_level_counts_0}
\title{Create the route level dataset (initial step)}
\usage{
create_route_level_counts_0(ebird, stop_level_data, taxonomy, config = config)
}
\arguments{
\item{stop_level_data}{result of `create_stop_level`}
}
\description{
Create the route level dataset (initial step)
}
