% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbbs.R
\name{create_stop_level_counts}
\alias{create_stop_level_counts}
\title{Main function to create stop level count dataset}
\usage{
create_stop_level_counts(ebird, taxonomy, config = config)
}
\description{
Main function to create stop level count dataset
}
