% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbbs.R
\name{create_stop_level_counts_0}
\alias{create_stop_level_counts_0}
\title{Create the stop level dataset (initial step)}
\usage{
create_stop_level_counts_0(ebird, taxonomy, config = config)
}
\arguments{
\item{ebird}{ebird dataset resulting from `get_ebird_data`}

\item{taxonomy}{taxonomy dataset}

\item{config}{mbbs configuration}
}
\description{
Create the stop level dataset (initial step)
}
