% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{create_stop_survey_list}
\alias{create_stop_survey_list}
\title{Name of function is temporary, should pick a better one
Returns a list of recording every stop+route+year combo that we
have data for. Also has locations of stop and if that has changed
between years when location data is available}
\usage{
create_stop_survey_list(locations, stop_level_counts)
}
\description{
Name of function is temporary, should pick a better one
Returns a list of recording every stop+route+year combo that we
have data for. Also has locations of stop and if that has changed
between years when location data is available
}
