% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_species_comments.R
\name{direct_to_sp_com_function}
\alias{direct_to_sp_com_function}
\title{Takes a row from stopsmbbs 
(an mbbs dataset that has gone through prepare_to_process_sp_com),
routes it to process the species_comment, 
and then returns the row with 's1:s20' filled in 
with the approprate counts}
\usage{
direct_to_sp_com_function(stopsmbbs_row)
}
\arguments{
\item{stopsmbbs_row}{a row from the mbbs that's gone through the prepare_to_process_sp_com 
function, ie: stopsmbbs. MUST have $species_comments, $count, and $s1:s20}
}
\description{
Takes a row from stopsmbbs 
(an mbbs dataset that has gone through prepare_to_process_sp_com),
routes it to process the species_comment, 
and then returns the row with 's1:s20' filled in 
with the approprate counts
}
