% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{distHaversine}
\alias{distHaversine}
\title{Compute distance between two points on the globe
taken from the 'geosphere' package, didn't want to add another dependency
ref: https://github.com/rspatial/geosphere/blob/master/R/distHaversine.R
modified lightly so p1 and p2 take c(lat, lon) which makes the most sense}
\usage{
distHaversine(p1, p2, r = 6378137)
}
\description{
Compute distance between two points on the globe
taken from the 'geosphere' package, didn't want to add another dependency
ref: https://github.com/rspatial/geosphere/blob/master/R/distHaversine.R
modified lightly so p1 and p2 take c(lat, lon) which makes the most sense
}
