% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\docType{data}
\name{ebird_cols}
\alias{ebird_cols}
\title{Column specification for ebird data}
\format{
An object of class \code{col_spec} of length 3.
}
\usage{
ebird_cols
}
\description{
Column specification for ebird data
}
\keyword{datasets}
