% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{ebird_import_checks}
\alias{ebird_import_checks}
\title{A basic set of import integrity checks
These do not check the validity of the data.}
\usage{
ebird_import_checks(dt)
}
\arguments{
\item{dt}{ebird data.frame}
}
\description{
A basic set of import integrity checks
These do not check the validity of the data.
}
\keyword{internal}
