% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{exclude_nonspecific_obs}
\alias{exclude_nonspecific_obs}
\title{Removes non specific observations}
\usage{
exclude_nonspecific_obs(ebird)
}
\arguments{
\item{ebird}{data.frame loaded from `load_ebird_data`}
}
\description{
Removes non specific observations
}
