% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{exclude_submissions}
\alias{exclude_submissions}
\title{Exclude submissions}
\usage{
exclude_submissions(ebird, exclusions)
}
\arguments{
\item{ebird}{data.frame loaded from `load_ebird_data`}

\item{exclusions}{character vector of submission ids to exclude}
}
\description{
Exclude submissions
}
