% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-level_excel.R
\name{extract_info_from_filename}
\alias{extract_info_from_filename}
\title{Pull mbbs_county, route_num, and year
out of a historical MBBS_XL
route_num MUST BE the first number in the filename
year MUST BE the second number in the filename}
\usage{
extract_info_from_filename(filename)
}
\arguments{
\item{filename}{a filename that contains the county, route num, and year}
}
\description{
Pull mbbs_county, route_num, and year
out of a historical MBBS_XL
route_num MUST BE the first number in the filename
year MUST BE the second number in the filename
}
