% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_comments.R
\name{extract_observers}
\alias{extract_observers}
\title{Extract observer from eBird comments}
\usage{
extract_observers(comments, ...)
}
\arguments{
\item{comments}{a `character` vector of eBird comments}

\item{...}{additional arguments passed to the `post` function}
}
\description{
Extract observer from eBird comments
}
