% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ebird_data.R
\name{extract_vehicles}
\alias{extract_vehicles}
\title{Extract vehicles from eBird comments}
\usage{
extract_vehicles(comments, ...)
}
\arguments{
\item{comments}{a `character` vector of MBBS eBird comments}

\item{...}{additional arguments passed to the `post` function}
}
\description{
Extract vehicles from eBird comments
}
