% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ebird_data.R
\name{extract_weather}
\alias{extract_weather}
\title{Extract notes from eBird comments}
\usage{
extract_weather(comments, ...)
}
\arguments{
\item{comments}{a `character` vector of MBBS eBird comments}

\item{...}{additional arguments passed to the `post` function}
}
\description{
Extract notes from eBird comments
}
