% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{filter_ebird_data}
\alias{filter_ebird_data}
\title{Removes subspecies, subgroup, or domestic type designations from the common
and scienfic name columns of an ebird csv}
\usage{
filter_ebird_data(ebird)
}
\arguments{
\item{ebird}{data.frame loaded from `load_ebird_data`}
}
\description{
Removes subspecies, subgroup, or domestic type designations from the common
and scienfic name columns of an ebird csv
}
