% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-level_obs_comments.R
\name{fix_obs_details}
\alias{fix_obs_details}
\title{Fix species comments
Fix unicode errors on enter signs,
and known typos within the dataset}
\usage{
fix_obs_details(x)
}
\arguments{
\item{x}{a character vector of observation details}
}
\description{
Fix species comments
Fix unicode errors on enter signs,
and known typos within the dataset
}
