% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-level_obs_comments.R
\name{fix_species_comments}
\alias{fix_species_comments}
\title{Fix species comments
Fix unicode errors on enter signs,
and known typos within the dataset}
\usage{
fix_species_comments(x)
}
\arguments{
\item{x}{a character vector of species comments}
}
\description{
Fix species comments
Fix unicode errors on enter signs,
and known typos within the dataset
}
