% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{get_ebird_data}
\alias{get_ebird_data}
\title{Gets the ebird data
This does *not* include:
* aligning common_names to standard taxonomy}
\usage{
get_ebird_data()
}
\description{
Gets the ebird data
This does *not* include:
* aligning common_names to standard taxonomy
}
