% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{get_ebird_locations}
\alias{get_ebird_locations}
\title{Gets the lat+lon of stop locations each year}
\usage{
get_ebird_locations(df)
}
\description{
Gets the lat+lon of stop locations each year
}
\keyword{internal}
