% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ebird_data.R
\name{get_exclusions}
\alias{get_exclusions}
\title{Get exclusions}
\usage{
get_exclusions(path = "inst/excluded_submissions.yml")
}
\arguments{
\item{path}{path/to/exclusions_map.yml}
}
\value{
a character vector of submission ids to exclude
}
\description{
Get exclusions
}
