% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_observers.R
\name{get_observer_quality}
\alias{get_observer_quality}
\title{Creates a fixed effect (numeric value) of observer quality, which reflects
!!!!! the standardized number of species the top observer in that year saw compared !!!EDIT
!!!!!!to the mean number of species seen on that route across all years !!!EDIT
observer_quality = max(obs1_quality, obs2_quality, obs3_quality, na.rm = TRUE)
Corrects for cases where a one-time observer accompanied a more experienced observer
and saw a high number of species in a particularly good year (putting their quality
above that of the more experienced observer)}
\usage{
get_observer_quality(mbbs_survey_events)
}
\arguments{
\item{mbbs_survey_events}{a dataframe with the list of survey events, importantly needs to include information about number of species and the observers for each survey}
}
\description{
Creates a fixed effect (numeric value) of observer quality, which reflects
!!!!! the standardized number of species the top observer in that year saw compared !!!EDIT
!!!!!!to the mean number of species seen on that route across all years !!!EDIT
observer_quality = max(obs1_quality, obs2_quality, obs3_quality, na.rm = TRUE)
Corrects for cases where a one-time observer accompanied a more experienced observer
and saw a high number of species in a particularly good year (putting their quality
above that of the more experienced observer)
}
