% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-level_transcribed.R
\name{get_stop_level_transcribed}
\alias{get_stop_level_transcribed}
\title{Get the stop-level data transcribed from paper files}
\usage{
get_stop_level_transcribed()
}
\description{
Get the stop-level data transcribed from paper files
}
