% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_stop_level_data.R
\name{granulate_to_stop}
\alias{granulate_to_stop}
\title{Replaces mbbs route-level data with the more granular stop-level information 
when available.
Works with any mbbs dataset, either the whole mbbs or one county.}
\usage{
granulate_to_stop(mbbs)
}
\arguments{
\item{mbbs}{Any mbbs dataset, either the whole survey area or one county}
}
\description{
Replaces mbbs route-level data with the more granular stop-level information 
when available.
Works with any mbbs dataset, either the whole mbbs or one county.
}
