% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{handle_deviations}
\alias{handle_deviations}
\title{Adds stops that were surveyed,
but no birds were observed.}
\usage{
handle_deviations(ebird, deviations)
}
\description{
Adds stops that were surveyed,
but no birds were observed.
}
