% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_historical_xls.R
\name{hist_xls_add_species_and_route_info}
\alias{hist_xls_add_species_and_route_info}
\title{Adds in the expected species and route info
To better match the usual processed mbbs data format}
\usage{
hist_xls_add_species_and_route_info(
  hist_xls,
  mbbs_survey_events = "data/mbbs_survey_events.rda"
)
}
\arguments{
\item{hist_xls}{a dataframe from a historical .xls}

\item{mbbs_survey_events}{rda of all surveys that have been run on the mbbs}
}
\value{
the hist_xls with added columns c(mbbs_county, route_num, year, observers, tax_order, sci_name)
}
\description{
Adds in the expected species and route info
To better match the usual processed mbbs data format
}
