% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_historical_xls.R
\name{hist_xls_change_xs_to_ones}
\alias{hist_xls_change_xs_to_ones}
\title{Converts 'x' and 'X' counts to 1
following method of mbbs route information
as compiled by Haven Wiley}
\usage{
hist_xls_change_xs_to_ones(hist_xls)
}
\description{
Converts 'x' and 'X' counts to 1
following method of mbbs route information
as compiled by Haven Wiley
}
