% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_historical_xls.R
\name{hist_xls_correct_common_names}
\alias{hist_xls_correct_common_names}
\title{Corrects species common_names to 
taxonomic standard}
\usage{
hist_xls_correct_common_names(x)
}
\arguments{
\item{x}{a dataframe from a historical .xls}
}
\description{
Corrects species common_names to 
taxonomic standard
}
