% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_historical_xls.R
\name{hist_xls_extract_county_num_year_from_filename}
\alias{hist_xls_extract_county_num_year_from_filename}
\title{Pull mbbs_county, route_num, and year
out of a historical MBBS_XL
route_num MUST BE the first number in the filename
year MUST BE the second number in the filename}
\usage{
hist_xls_extract_county_num_year_from_filename(filename)
}
\arguments{
\item{filename}{a filename that contains the county, route num, and year}
}
\description{
Pull mbbs_county, route_num, and year
out of a historical MBBS_XL
route_num MUST BE the first number in the filename
year MUST BE the second number in the filename
}
