% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_historical_xls.R
\name{hist_xls_filter_to_species_code_rows}
\alias{hist_xls_filter_to_species_code_rows}
\title{Filters to only rows where the first column is a species code
that is four capital letters
or acsp - "Accipiter sp."}
\usage{
hist_xls_filter_to_species_code_rows(x)
}
\arguments{
\item{x}{a dataframe from a historical .xls}
}
\description{
Filters to only rows where the first column is a species code
that is four capital letters
or acsp - "Accipiter sp."
}
