% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_historical_xls.R
\name{hist_xls_flag_missed_species}
\alias{hist_xls_flag_missed_species}
\title{Flags if a species code has been missed
Used to ensure there are no errors with any new .xls sheets that come in}
\usage{
hist_xls_flag_missed_species(df)
}
\arguments{
\item{df}{a dataframe that contains the column "sequence"}
}
\description{
Flags if a species code has been missed
Used to ensure there are no errors with any new .xls sheets that come in
}
