% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_historical_xls.R
\name{hist_xls_match_mbbs_format}
\alias{hist_xls_match_mbbs_format}
\title{Match MMBS format
This doesn't add every mbbs column
But fills in those for species or route info
that are expected}
\usage{
hist_xls_match_mbbs_format(
  hist_xls,
  mbbs_survey_events = "data/mbbs_survey_events.rda"
)
}
\description{
Match MMBS format
This doesn't add every mbbs column
But fills in those for species or route info
that are expected
}
