% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_historical_xls.R
\name{hist_xls_purrr_directories}
\alias{hist_xls_purrr_directories}
\title{Takes a file directory that contains
Folders of historical .xls
and processes them all into one df}
\usage{
hist_xls_purrr_directories(directory = "inst/extdata/stop_level_data")
}
\description{
Takes a file directory that contains
Folders of historical .xls
and processes them all into one df
}
