% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_historical_xls.R
\name{hist_xls_purrr_files_in_directory}
\alias{hist_xls_purrr_files_in_directory}
\title{Takes a file directory of historical .xls
files, and processes them,
returning a dataframe that contains
all the information from that directory's files
broken down to the stop level.}
\usage{
hist_xls_purrr_files_in_directory(
  directory = "inst/extdata/stop_level_data/Pippen_Durham_stops"
)
}
\description{
Takes a file directory of historical .xls
files, and processes them,
returning a dataframe that contains
all the information from that directory's files
broken down to the stop level.
}
