% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_historical_xls.R
\name{hist_xls_rename_columns}
\alias{hist_xls_rename_columns}
\title{Rename the columns of a historical xls}
\usage{
hist_xls_rename_columns(x)
}
\arguments{
\item{x}{a dataframe from a historical .xls}
}
\description{
Rename the columns of a historical xls
}
