% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_historical_xls.R
\name{hist_xls_run_checks}
\alias{hist_xls_run_checks}
\title{A set of integrity checks
These do not check the validity of the data.}
\usage{
hist_xls_run_checks(df)
}
\arguments{
\item{df}{a df from an .xls that has just been processed through
the rest of the hist_xls_process_xls function}
}
\description{
A set of integrity checks
These do not check the validity of the data.
}
