% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_historical_xls.R
\name{hist_xls_sci_name_check}
\alias{hist_xls_sci_name_check}
\title{Defined separately from run_checks for testing purposes
Tests that all rows except for where common_name is
'Accipiter species' have a non-NA sci_name column}
\usage{
hist_xls_sci_name_check(df)
}
\arguments{
\item{df}{a df from an .xls that has just been processed through
the rest of the hist_xls_process_xls function}
}
\description{
Defined separately from run_checks for testing purposes
Tests that all rows except for where common_name is
'Accipiter species' have a non-NA sci_name column
}
