% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical.R
\docType{data}
\name{historical_cols}
\alias{historical_cols}
\title{Column specification for historical data}
\format{
An object of class \code{character} of length 15.
}
\usage{
historical_cols
}
\description{
Column specification for historical data
}
\keyword{datasets}
