% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ebird_data.R
\name{import_ebird_data}
\alias{import_ebird_data}
\title{Import an ebird export into R}
\usage{
import_ebird_data(path, run_checks = TRUE)
}
\arguments{
\item{path}{path/to/ebird_export.csv}
}
\description{
Import an ebird export into R
}
