% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{list_ebird_files}
\alias{list_ebird_files}
\title{Create a data.frame listing available eBird files}
\usage{
list_ebird_files(path = config$ebird_data_dir)
}
\description{
Create a data.frame listing available eBird files
}
