% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-level_excel.R
\name{list_stop_level_xls_files}
\alias{list_stop_level_xls_files}
\title{Takes a file directory that contains
Folders of historical .xls
and processes them all into one df}
\usage{
list_stop_level_xls_files(directory = config$stop_level_xls_dir)
}
\arguments{
\item{directory}{directory containing containing historical xls files.}
}
\value{
vector of filenames of stop level files
}
\description{
Takes a file directory that contains
Folders of historical .xls
and processes them all into one df
}
