% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{load_ebird_data}
\alias{load_ebird_data}
\title{Load the *most recent* ebird export per county
(i.e. MBBS ebird account)
into an R `data.frame`.
according the `ebird_cols` specification.
Additionally renames the variables.}
\usage{
load_ebird_data()
}
\description{
NOTE: readr::read_csv gives warnings upon parsing the ebird CSVs.
      As best I (BS) can tell as of 20241001,
      the problems stem from non-regularity in the ebird exports.
      Sometimes, columns are missing from a row.
      In other words,
      the ebird csv can have different numbers of columns per row!
      This issue does not seem to affect our data,
      but this is something to review.
}
