% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical.R
\name{load_historical_data}
\alias{load_historical_data}
\title{Loads the historical (pre-eBird) MBBS data from disk}
\usage{
load_historical_data(path = config$historical_data_dir)
}
\description{
This function simply loads the csv files
in the `historical` data directory
as a `data.frame`.
}
