% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{locations_workflow}
\alias{locations_workflow}
\title{get data from the $stop_counts
left join the ebird locations data
called in ebird_import}
\usage{
locations_workflow(df)
}
\description{
get data from the $stop_counts
left join the ebird locations data
called in ebird_import
}
