% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbbs_data.R
\docType{data}
\name{mbbs}
\alias{mbbs}
\title{Data from the Mini-Bird Breeding Survey}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 45016 rows and 35 columns.
}
\usage{
mbbs
}
\description{
TODO: Describe the data
}
\keyword{datasets}
