% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbbs_data.R
\docType{data}
\name{mbbs_chatham}
\alias{mbbs_chatham}
\title{Data from the Mini-Bird Breeding Survey}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 19841 rows and 33 columns.
}
\usage{
mbbs_chatham
}
\description{
TODO: Describe the data
}
\keyword{datasets}
