% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mbbs_data.R
\name{mbbs_generate_route_ID}
\alias{mbbs_generate_route_ID}
\title{Add route_ID column}
\usage{
mbbs_generate_route_ID(x)
}
\arguments{
\item{x}{the output of combine_site_ebird}
}
\description{
Add route_ID column
}
