% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbbs_data.R
\docType{data}
\name{mbbs_orange}
\alias{mbbs_orange}
\title{Data from the Mini-Bird Breeding Survey}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17705 rows and 34 columns.
}
\usage{
mbbs_orange
}
\description{
TODO: Describe the data
}
\keyword{datasets}
