% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbbs_data.R
\docType{data}
\name{mbbs_results}
\alias{mbbs_results}
\title{MBBS results data for use in data displays in other websites}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74 rows and 19 columns.}
\usage{
mbbs_results
}
\description{
TODO: Describe the data
}
\keyword{datasets}
