% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbbs_data.R
\docType{data}
\name{mbbs_routes}
\alias{mbbs_routes}
\title{Data.frame containing information about the routes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34 rows and 4 columns.
}
\usage{
mbbs_routes
}
\description{
TODO: Describe the data
}
\keyword{datasets}
