% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbbs_data.R
\docType{data}
\name{mbbs_species}
\alias{mbbs_species}
\title{Additional Data about each species in the Mini-Bird Breeding Survey}
\format{An object of class \code{data.frame} with 115 rows and 6 columns.}
\usage{
mbbs_species
}
\description{
TODO: Describe the data
}
\keyword{datasets}
