% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbbs_data.R
\docType{data}
\name{mbbs_survey_events}
\alias{mbbs_survey_events}
\title{Data from the Mini-Bird Breeding Survey}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 751 rows and 22 columns.
}
\usage{
mbbs_survey_events
}
\description{
TODO: Describe the data
}
\keyword{datasets}
