% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_species_comments.R
\name{pad_or_truncate}
\alias{pad_or_truncate}
\title{takes a list and either extends it with 0s to the max_length or cuts it down}
\usage{
pad_or_truncate(x, max_length = 20)
}
\description{
takes a list and either extends it with 0s to the max_length or cuts it down
}
