% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_comments.R
\name{parse_comments}
\alias{parse_comments}
\title{Parse all stop-level comments}
\usage{
parse_comments(x)
}
\arguments{
\item{x}{list of `ebird` comments}
}
\description{
Parse all stop-level comments
}
