% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{parse_common_name}
\alias{parse_common_name}
\title{Parse the character common_name vector}
\usage{
parse_common_name(common_name)
}
\arguments{
\item{common_name}{character vector of ebird common names}
}
\description{
Parse the character common_name vector
}
