% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{parse_count}
\alias{parse_count}
\title{Parse the character count vector into an integer vector
NOTE: This parser maps the ebird "X" (meant to indicate presence)
      into 1.
      As as 20241001 there are only 11 observations that use "X",
      the most recent in 2012;
      most of the species in this case are reasonably uncommon.}
\usage{
parse_count(count)
}
\arguments{
\item{count}{character vector of ebird counts}
}
\description{
Parse the character count vector into an integer vector
NOTE: This parser maps the ebird "X" (meant to indicate presence)
      into 1.
      As as 20241001 there are only 11 observations that use "X",
      the most recent in 2012;
      most of the species in this case are reasonably uncommon.
}
