% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{parse_county}
\alias{parse_county}
\title{Parse a route's county from the ebird location name}
\usage{
parse_county(location)
}
\arguments{
\item{location}{character vector of ebird locations}
}
\description{
Parse a route's county from the ebird location name
}
