% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-level_obs_comments.R
\name{parse_lengtheq20}
\alias{parse_lengtheq20}
\title{When obs_details splits into a vector of length 20,
we assume each position corresponds to a stop (in the correct order)
and extract the count accordingly.}
\usage{
parse_lengtheq20(x)
}
\description{
When obs_details splits into a vector of length 20,
we assume each position corresponds to a stop (in the correct order)
and extract the count accordingly.
}
