% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-level_obs_comments.R
\name{parse_lengthlt20}
\alias{parse_lengthlt20}
\title{When obs_details splits into a vector of length less than 20
(after `pad_or_truncate`),
they are 2 cases to handle:
(a) the number in the field corresponds to the *only* stop
    at which the count was observed;
    in this case, we obtain the count from the route level data.
(b) a pattern of i=#,j=#, etc,
    where the LHS of "=" the stop numbers
    and the RHS is the count}
\usage{
parse_lengthlt20(x, count, subid)
}
\description{
When obs_details splits into a vector of length less than 20
(after `pad_or_truncate`),
they are 2 cases to handle:
(a) the number in the field corresponds to the *only* stop
    at which the count was observed;
    in this case, we obtain the count from the route level data.
(b) a pattern of i=#,j=#, etc,
    where the LHS of "=" the stop numbers
    and the RHS is the count
}
