% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{parse_route_num}
\alias{parse_route_num}
\title{Parse a route's number from the ebird location name}
\usage{
parse_route_num(location)
}
\arguments{
\item{location}{character vector of ebird locations}
}
\description{
Parse a route's number from the ebird location name
}
