% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{parse_sci_name}
\alias{parse_sci_name}
\title{Parse the character sci_name vector}
\usage{
parse_sci_name(sci_name)
}
\arguments{
\item{sci_name}{character vector of ebird scientific names}
}
\description{
Parse the character sci_name vector
}
