% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_comments.R
\name{parse_single_comment}
\alias{parse_single_comment}
\title{Parse a *single* submission comment}
\usage{
parse_single_comment(x)
}
\arguments{
\item{x}{character vector of `ebird` of *single* submission's comments}
}
\description{
NOTE: This only does an *initial* parsing of habitat data;
      i.e. it only extracts data to the right of (e.g.) "habitat==".
}
