% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{parse_stop_num}
\alias{parse_stop_num}
\title{Parse a stop number from the ebird location name}
\usage{
parse_stop_num(location)
}
\arguments{
\item{location}{character vector of ebird locations}
}
\description{
Parse a stop number from the ebird location name
}
