% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_comments.R
\name{postprocess_comments}
\alias{postprocess_comments}
\title{PostProcess comments}
\usage{
postprocess_comments(x)
}
\arguments{
\item{x}{list of list(vehicle, ...) obtained after `parse_comments`}
}
\description{
PostProcess comments
}
