% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-level_obs_comments.R
\name{prep_obs_details_data}
\alias{prep_obs_details_data}
\title{Prepare ebird dataset for processing observation comments}
\usage{
prep_obs_details_data(ebird)
}
\arguments{
\item{ebird}{ebird data}
}
\description{
Prepare ebird dataset for processing observation comments
}
