% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mbbs_data.R
\name{prepare_mbbs_data}
\alias{prepare_mbbs_data}
\title{A function that combines and aligns MBBS data scraped from the old website
with ebird data.}
\usage{
prepare_mbbs_data(ebird_dt, mbbs_site_dt, ebird_taxonomy)
}
\arguments{
\item{ebird_dt}{a `data.frame` imported using `import_ebird_data` for a single
county}

\item{mbbs_site_dt}{a `data.frame` containing MBBS data scraped from the old
website for a single county.}

\item{ebird_taxonomy}{a `data.frame` containing common and scientific names
of ebird taxonomy}
}
\value{
a `list` containing `site` (processed "website" data) and `ebird`
   (processed ebird data)
}
\description{
This is one big function that does a lot of data munging!
}
