% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-level_obs_comments.R
\name{process_obs_details}
\alias{process_obs_details}
\title{Process species comments
Takes data/mbbs.rda rows that have non NA species_comments
and processes the row from a full-route summary of count for each species
to break it out by stop_num.}
\usage{
process_obs_details(ebird)
}
\arguments{
\item{ebird}{ebird data}
}
\value{
a df where each row is a count of a species at a stop
}
\description{
Process species comments
Takes data/mbbs.rda rows that have non NA species_comments
and processes the row from a full-route summary of count for each species
to break it out by stop_num.
}
