% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_observers.R
\name{process_observers}
\alias{process_observers}
\title{Full workflow for processing observers}
\usage{
process_observers(mbbs_county, county)
}
\arguments{
\item{mbbs_county}{mbbs data.frame}

\item{county}{which county is being processed, orange, durham, or chatham lower case}
}
\value{
mbbs_county
}
\description{
Full workflow for processing observers
}
