% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_stop_level_data.R
\name{process_species_comments}
\alias{process_species_comments}
\title{Processes the stop level information left in the species_comments column of mbbs rows imported from ebird
Returns df of mbbs data that had an informative species_comments broken down to the stop level}
\usage{
process_species_comments(mbbs)
}
\arguments{
\item{mbbs}{Any mbbs dataset, either the whole survey area or one county}
}
\description{
Processes the stop level information left in the species_comments column of mbbs rows imported from ebird
Returns df of mbbs data that had an informative species_comments broken down to the stop level
}
