% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-level_excel.R
\name{process_stop_level_xls}
\alias{process_stop_level_xls}
\title{Process stop-level excel sheets
They should all have the same format,
if it is in the wrong format
there will be errors}
\usage{
process_stop_level_xls(file)
}
\arguments{
\item{file}{name of xls file to process}
}
\value{
hist_xls each row is a species with columns s1:s20
   containing the count at each stop
}
\description{
Process stop-level excel sheets
They should all have the same format,
if it is in the wrong format
there will be errors
}
