% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_observers.R
\name{propogate_observers_across_stops}
\alias{propogate_observers_across_stops}
\title{Function to add observers to route stops 2:20}
\usage{
propogate_observers_across_stops(mbbs_county)
}
\arguments{
\item{mbbs_county}{mbbs data.frame}
}
\value{
mbbs_county input dataframe with the observer column added to all stops
}
\description{
Function to add observers to route stops 2:20
}
