% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-level_transcribed.R
\name{read_transcribed_csv}
\alias{read_transcribed_csv}
\title{Reads in a transcribed file csv and adds the name of the transcriber}
\usage{
read_transcribed_csv(
  filepath = "inst/stop-level/transcribed_paper_files/transcribed_paper_files_GEL.csv"
)
}
\arguments{
\item{filepath}{file path to the file to be read in}
}
\value{
transcribed the df read in from the csv from the filepath
}
\description{
Reads in a transcribed file csv and adds the name of the transcriber
}
