% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ebird_data.R
\name{rename_ebird_data}
\alias{rename_ebird_data}
\title{Renames and selects columns from data.frame created from the ebird csv export.}
\usage{
rename_ebird_data(dt)
}
\arguments{
\item{dt}{a data.frame imported from an ebird export.}
}
\description{
Renames and selects columns from data.frame created from the ebird csv export.
}
