% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ebird_data.R
\name{rename_subspecies}
\alias{rename_subspecies}
\title{Removes subspecies, subgroup, or domestic type designations from the common
and scienfic name columns of an ebird csv}
\usage{
rename_subspecies(dt)
}
\arguments{
\item{dt}{data table / data frame}
}
\description{
Removes subspecies, subgroup, or domestic type designations from the common
and scienfic name columns of an ebird csv
}
