% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_observers.R
\name{save_observer_table}
\alias{save_observer_table}
\title{Arranges and saves a new version of the observer table}
\usage{
save_observer_table(
  observer_table,
  file = "inst/extdata/observer_conversion_table.csv"
)
}
\arguments{
\item{observer_table}{observer table data.frame}
}
\description{
Arranges and saves a new version of the observer table
}
