% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_species_comments.R
\name{sp_com_comma_seperated}
\alias{sp_com_comma_seperated}
\title{Use when species comments follows ,,,,,,1,,,, format
eg. where commas seperate counts at each stop. 
Takes a string and returns a list of 20 numbers}
\usage{
sp_com_comma_seperated(x)
}
\arguments{
\item{x}{a string from the species_comment column formatted with commas}
}
\description{
Use when species comments follows ,,,,,,1,,,, format
eg. where commas seperate counts at each stop. 
Takes a string and returns a list of 20 numbers
}
