% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_species_comments.R
\name{sp_com_only_stop}
\alias{sp_com_only_stop}
\title{Use when species comment is ie: "Stop 6"
The count for that stop is the count for the whole route
and the species was not seen at any other stops}
\usage{
sp_com_only_stop(x, count = -999)
}
\arguments{
\item{x}{a string from the species_comment column formatted by "stop=count"}

\item{count}{an integer, from the count column of the mbbs}
}
\description{
Use when species comment is ie: "Stop 6"
The count for that stop is the count for the whole route
and the species was not seen at any other stops
}
