% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_species_comments.R
\name{sp_com_stop_equals_count}
\alias{sp_com_stop_equals_count}
\title{Use when species comments follows 1=5 or st 1=5 to stop 1=5 format
Takes a string and returns a list of 20 numbers}
\usage{
sp_com_stop_equals_count(x)
}
\arguments{
\item{x}{a string from the species_comment column formatted by "stop=count"}
}
\value{
out a standardized list of 20 integers, representing the 
   species count at 20 stops.
}
\description{
Use when species comments follows 1=5 or st 1=5 to stop 1=5 format
Takes a string and returns a list of 20 numbers
}
