% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_comments.R
\name{split_comments}
\alias{split_comments}
\title{Split comment ebird comments at ";"}
\usage{
split_comments(comments)
}
\arguments{
\item{comments}{`ebird` comments}
}
\description{
Split comment ebird comments at ";"
}
