% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_observers.R
\name{standardize_observer_table}
\alias{standardize_observer_table}
\title{Function to convert obs1 obs2 and obs3 from the main_observer_table
to the standard format using mini_observer_table conversions
Also creates the standardized observers column.}
\usage{
standardize_observer_table(save = TRUE)
}
\arguments{
\item{save}{TRUE if the updated table should be saved. Set to FALSE when testing}
}
\description{
Function to convert obs1 obs2 and obs3 from the main_observer_table
to the standard format using mini_observer_table conversions
Also creates the standardized observers column.
}
