% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-level_excel.R
\name{stop_level_xls_checks}
\alias{stop_level_xls_checks}
\title{A set of integrity checks
These do not check the validity of the data.}
\usage{
stop_level_xls_checks(df)
}
\arguments{
\item{df}{a df from an .xls that has just been processed through
the rest of the hist_xls_process_xls function}
}
\description{
A set of integrity checks
These do not check the validity of the data.
}
