% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{stops_moved}
\alias{stops_moved}
\title{Identify stops that were moved in the list of deviations.}
\usage{
stops_moved(deviations)
}
\value{
a data.frame of year/county/route/stop_num which need to be added
        to the ebird data
}
\description{
Identify stops that were moved in the list of deviations.
}
