% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{stops_to_include}
\alias{stops_to_include}
\title{Identify stops to include from the list of deviations.}
\usage{
stops_to_include(deviations)
}
\value{
a data.frame of year/county/route/stop_num which need to be added
        to the ebird data
}
\description{
Identify stops to include from the list of deviations.
}
