% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_transcribed_paper.R
\name{tpaper_find_mismatches}
\alias{tpaper_find_mismatches}
\title{Finds mismatches between dataframes, and adds the source as a column}
\usage{
tpaper_find_mismatches(
  df1 =
    read.csv("inst/stop-level/transcribed_paper_files/transcribed_paper_files_GEL.csv"),
  df2 =
    read.csv("inst/stop-level/transcribed_paper_files/transcribed_paper_files_ADS.csv"),
  source1 = "GEL",
  source2 = "ADS",
  overlap = read.csv("inst/stop-level/transcribed_paper_files/mbbs_paper_files_list.csv")
)
}
\arguments{
\item{df1}{first transcribed paper df}

\item{df2}{second transcribed paper df}

\item{source1}{character}

\item{source2}{character}

\item{overlap}{the list of who transcribed which files}
}
\value{
mismatches -df of ALL mismatches between df1 and df2
}
\description{
Finds mismatches between dataframes, and adds the source as a column
}
