% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-level_transcribed.R
\name{tpaper_find_overlap}
\alias{tpaper_find_overlap}
\title{Returns the rows from df1 and df2 that had double entry}
\usage{
tpaper_find_overlap(
  df1 =
    read.csv("inst/stop-level/transcribed_paper_files/transcribed_paper_files_GEL.csv"),
  df2 =
    read.csv("inst/stop-level/transcribed_paper_files/transcribed_paper_files_ADS.csv"),
  source1 = "GEL",
  source2 = "ADS",
  overlap = read.csv(config$transcribed_paper_list)
)
}
\arguments{
\item{df1}{first transcribed paper df}

\item{df2}{second transcribed paper df}

\item{source1}{character}

\item{source2}{character}

\item{overlap}{the list of who transcribed which files}
}
\value{
x df with only the rows from df1 and df2 that had double entry
}
\description{
Returns the rows from df1 and df2 that had double entry
}
