% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{transform_ebird_data}
\alias{transform_ebird_data}
\title{Transformation of the ebird data to conform to format
downstream consumers expect}
\usage{
transform_ebird_data(ebird)
}
\description{
Transformation of the ebird data to conform to format
downstream consumers expect
}
