% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical.R
\name{transform_historical_data}
\alias{transform_historical_data}
\title{Transform historical (pre-eBird) data}
\usage{
transform_historical_data(historical)
}
\description{
Transform historical (pre-eBird) data
}
