% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_observers.R
\name{update_observer_tables}
\alias{update_observer_tables}
\title{Updates the main observer table when new route
+ observer combos are present
calls update_mini_observer_table
and then updates the main table based on the mini table.
Called from import_data as part of process_obervers() during the processing of new data}
\usage{
update_observer_tables(mbbs_county, selected_county)
}
\arguments{
\item{mbbs_county}{mbbs data.frame, must end in and underscore then the name of the county ie: _durham, _orange, _chatham}

\item{selected_county}{county that the observer table should be filtered on}
}
\description{
Updates the main observer table when new route
+ observer combos are present
calls update_mini_observer_table
and then updates the main table based on the mini table.
Called from import_data as part of process_obervers() during the processing of new data
}
