% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-events.R
\name{update_survey_events}
\alias{update_survey_events}
\title{Updates survey_events if needed by rbinding new and updated surveys,
  then updates survey_events
  Called from import_data after updating the data/mbbs.rda files
  Not part of the process_observers workflow.}
\usage{
update_survey_events(path = config$survey_list)
}
\description{
Updates survey_events if needed by rbinding new and updated surveys,
  then updates survey_events
  Called from import_data after updating the data/mbbs.rda files
  Not part of the process_observers workflow.
}
