% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-list.R
\name{update_survey_list}
\alias{update_survey_list}
\title{Updates survey_list if needed by:
  finding any ebird surveys not yet on the survey_list
  standardizing obs1/obs2/obs3
  adding any new observer matches to the observer_conversion_table,
  taking input on new names,
  and then adding the now observer-standardized new surveys to survey_list}
\usage{
update_survey_list(ebird, config, path = config$survey_list, save = TRUE)
}
\description{
Updates survey_list if needed by:
  finding any ebird surveys not yet on the survey_list
  standardizing obs1/obs2/obs3
  adding any new observer matches to the observer_conversion_table,
  taking input on new names,
  and then adding the now observer-standardized new surveys to survey_list
}
