% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-level_transcribed.R
\name{update_transcribed_paper}
\alias{update_transcribed_paper}
\title{Update transcribed paper
Reads in the transcribed files
Saves any mismatches
Assuming no mismatches, adds together all the double-entry data
writes the stop-level-transcribed-paper csv}
\usage{
update_transcribed_paper()
}
\value{
transcribed_paper df of all the observations transcribed from the paper files,
 returned just in case this function is called to set a variable.
}
\description{
Update transcribed paper
Reads in the transcribed files
Saves any mismatches
Assuming no mismatches, adds together all the double-entry data
writes the stop-level-transcribed-paper csv
}
