% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_date_range}
\alias{validate_date_range}
\title{Check that dates are within the study dates}
\usage{
validate_date_range(x)
}
\arguments{
\item{x}{a date vector}
}
\value{
a logical vector indicating that the date is between 5/15 and 6/30
}
\description{
Check that dates are within the study dates
}
