% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_comments.R
\name{clean_comments}
\alias{clean_comments}
\title{Clean up the eBird}
\usage{
clean_comments(comments)
}
\arguments{
\item{comments}{a `character` vector of MBBS eBird comments}
}
\description{
Clean up the eBird
}
