% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical.R
\name{clean_common_names}
\alias{clean_common_names}
\title{Clean the species names of the data scraped from the old MBBS website}
\usage{
clean_common_names(x)
}
\arguments{
\item{x}{character vector of common names}
}
\description{
Clean the species names of the data scraped from the old MBBS website
}
