% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_comments.R
\name{comment_workflow}
\alias{comment_workflow}
\title{Full workflow for processing comments}
\usage{
comment_workflow(eBird_dt)
}
\arguments{
\item{eBird_dt}{a `data.frame` of imported eBird counts}
}
\value{
a `data.frame` with one row per submission ID in `eBird_dt`
}
\description{
Full workflow for processing comments
}
