% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ebird_data.R
\name{exclude_submissions}
\alias{exclude_submissions}
\title{Exclude submissions}
\usage{
exclude_submissions(dt, exclusions)
}
\arguments{
\item{dt}{ebird `data.frame`}

\item{exclusions}{character vector of submission ids to exclude}
}
\value{
a `data.frame` without the exclusions
}
\description{
Exclude submissions
}
