% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_comments.R
\name{extract_vehicles}
\alias{extract_vehicles}
\title{Extract vehicles from eBird comments}
\usage{
extract_vehicles(comments, ...)
}
\arguments{
\item{comments}{a `character` vector of eBird comments}

\item{...}{additional arguments passed to the `post` function}
}
\description{
Extract vehicles from eBird comments
}
