% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ebird_data.R
\name{filter_ebird_data}
\alias{filter_ebird_data}
\title{Filters ebird csv export.}
\usage{
filter_ebird_data(dt)
}
\arguments{
\item{dt}{a data.frame imported from an ebird export.}
}
\description{
Filters ebird csv export.
}
