% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_import.R
\name{get_exclusions}
\alias{get_exclusions}
\title{Get exclusions}
\usage{
get_exclusions(file = config$excluded_submissions)
}
\value{
a character vector of submission ids to exclude
}
\description{
Get exclusions
}
