% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_historical_xls.R
\name{hist_xls_process_xls}
\alias{hist_xls_process_xls}
\title{Process historical excel sheets
They should all have the same format,
if it is in the wrong format
there will be errors}
\usage{
hist_xls_process_xls(filenames)
}
\arguments{
\item{filenames}{a list of historical mbbs .xls files within a folder}
}
\description{
Process historical excel sheets
They should all have the same format,
if it is in the wrong format
there will be errors
}
