% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbbs_data.R
\docType{data}
\name{mbbs}
\alias{mbbs}
\title{Data.frame containing information about the routes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51121 rows and 35 columns.
}
\usage{
mbbs
}
\description{
TODO: Describe the data
}
\keyword{datasets}
