% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbbs_data.R
\docType{data}
\name{mbbs}
\alias{mbbs}
\title{Data from the Mini-Bird Breeding Survey}
\format{An object of class \code{data.frame} with 13169 rows and 29 columns.}
\usage{
mbbs
}
\description{
TODO: Describe the data
}
\keyword{datasets}
