% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_observers.R
\name{observers_extractor}
\alias{observers_extractor}
\title{Pull out observers from the checklist_comments column that process_comments missed}
\usage{
observers_extractor(mbbs_county)
}
\arguments{
\item{mbbs_county}{mbbs data.frame}
}
\value{
mbbs_county the input dataframe with fixed observers
}
\description{
Pull out observers from the checklist_comments column that process_comments missed
}
