% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_species_comments.R
\name{pad_or_truncate}
\alias{pad_or_truncate}
\title{takes a list and either extends it with 0s to the max_length or cuts it down}
\usage{
pad_or_truncate(x, max_length = 20)
}
\arguments{
\item{x}{a list of numbers}

\item{max_length}{the final length that list x should be.}
}
\value{
x either cut down to the max_length or with 0's added to the end to
 get it to the max_length
}
\description{
takes a list and either extends it with 0s to the max_length or cuts it down
}
