% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ebird_data.R
\name{postprocess_comments}
\alias{postprocess_comments}
\title{Workflow for postprocessing eBird comments}
\usage{
postprocess_comments(comments)
}
\arguments{
\item{comments}{a `character` vector of MBBS eBird comments}
}
\description{
Workflow for postprocessing eBird comments
}
