% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_comments.R
\name{postprocess_comments}
\alias{postprocess_comments}
\title{Workflow for postprocessing eBird comments}
\usage{
postprocess_comments(comments)
}
\arguments{
\item{comments}{a `character` vector of eBird comments}
}
\description{
Workflow for postprocessing eBird comments
}
