% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_species_comments.R
\name{prepare_to_process_sp_com}
\alias{prepare_to_process_sp_com}
\title{Prepare mbbs dataset for processing species comments}
\usage{
prepare_to_process_sp_com(mbbs)
}
\description{
Prepare mbbs dataset for processing species comments
}
