% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_comments.R
\name{preprocess_comments}
\alias{preprocess_comments}
\title{Workflow for preprocessing eBird comments}
\usage{
preprocess_comments(comments)
}
\arguments{
\item{comments}{a `character` vector of eBird comments}
}
\description{
Workflow for preprocessing eBird comments
}
