% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_comments.R
\name{process_comments}
\alias{process_comments}
\title{Workflow for processing eBird comments}
\usage{
process_comments(comments)
}
\arguments{
\item{comments}{a `character` vector of eBird comments}
}
\description{
Workflow for processing eBird comments
}
