% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ebird_data.R
\name{process_comments}
\alias{process_comments}
\title{Workflow for processing eBird comments}
\usage{
process_comments(comments)
}
\arguments{
\item{comments}{a `character` vector of MBBS eBird comments}
}
\description{
Workflow for processing eBird comments
}
