% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_observers.R
\name{save_observer_table}
\alias{save_observer_table}
\title{Arranges and saves a new version of the observer table}
\usage{
save_observer_table(
  observer_table,
  file = config$main_observer_conversion_table
)
}
\arguments{
\item{observer_table}{observer table data.frame}

\item{file}{where the observer table is being saved}
}
\description{
Arranges and saves a new version of the observer table
}
