% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_observers.R
\name{update_mini_observer_table}
\alias{update_mini_observer_table}
\title{Add any new entries to the mini_observer_conversion_table}
\usage{
update_mini_observer_table(
  observer_table = read.csv(config$main_observer_conversion_table, header = TRUE),
  mini_observer_table = read.csv(config$mini_observer_conversion_table, header = TRUE),
  save = TRUE
)
}
\arguments{
\item{observer_table}{the main observer conversion table. Taken as an argument so the most recent version (potentially one in progress) can be used.}

\item{mini_observer_table}{the mini observer table. Taken as an argument so we can just load it already rather than within the function.}

\item{save}{TRUE if the updated table should be saved. Set to FALSE when testing}
}
\value{
an updated version of the main observer conversion table
}
\description{
Add any new entries to the mini_observer_conversion_table
}
