% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_observers.R
\name{update_mini_observer_table}
\alias{update_mini_observer_table}
\title{Add new entries to the mini_observer_conversion_table}
\usage{
update_mini_observer_table()
}
\description{
Add new entries to the mini_observer_conversion_table
}
