% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_observers.R
\name{update_mini_observer_table}
\alias{update_mini_observer_table}
\title{Add any new entries to the mini_observer_conversion_table}
\usage{
update_mini_observer_table(
  observer_table = read.csv("inst/extdata/main_observer_conversion_table.csv", header =
    TRUE)
)
}
\arguments{
\item{observer_table}{the main observer conversion table. Taken as an argument so the most recent version (potentially one in progress) can be used.}
}
\value{
an updated version of the main observer conversion table
}
\description{
Add any new entries to the mini_observer_conversion_table
}
