% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_observers.R
\name{update_observer_table}
\alias{update_observer_table}
\title{Interactive program to update the observer table when new route 
+ observer combos are present}
\usage{
update_observer_table(mbbs_county, selected_county)
}
\arguments{
\item{mbbs_county}{mbbs data.frame, must end in and underscore then the name of the county ie: _durham, _orange, _chatham}

\item{selected_county}{county that the observer table should be filtered on}
}
\description{
Interactive program to update the observer table when new route 
+ observer combos are present
}
