% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-list.R
\name{update_observer_table}
\alias{update_observer_table}
\title{Add any new entries to the observer_conversion_table
Does not return anything.}
\usage{
update_observer_table(ebird_surveys, config, save = TRUE)
}
\description{
Add any new entries to the observer_conversion_table
Does not return anything.
}
